+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/accountinfo/ScheduledRelease.java
@@ -1,9 +1,8 @@
 package com.concordium.sdk.responses.accountinfo;
 
 import com.concordium.sdk.transactions.CCDAmount;
-import com.concordium.sdk.types.Timestamp;
-
 import com.concordium.sdk.transactions.Hash;
+import com.concordium.sdk.types.Timestamp;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -11,11 +10,26 @@ import lombok.ToString;
 
 import java.util.List;
 
+/**
+ * A CCD release in the future.
+ */
 @Getter
 @ToString
 public final class ScheduledRelease {
+
+    /**
+     * The time when the CCD will be released for the account.
+     */
     private final Timestamp timestamp;
+
+    /**
+     * The amount that will be released.
+     */
     private final CCDAmount amount;
+
+    /**
+     * A list of 'origin' transactions for the {@link ScheduledRelease}
+     */
     private final List<Hash> transactions;
 
     @JsonCreator